Update massFileDownloader.js
Just added a lightning-textarea with the URL removing the window.alert(). The previous one wasn't showing the whole URL if needed and now the URL is also getting updated according to the file selection (checkbox selection).

On the JS, I've changed it a little bit to create a downloadString variable, as well as getters/setters to update the url dynamically.